+++ Cargo.toml
@@ -4,3 +4,9 @@ members = ["polars-rows-iter", "polars-rows-iter-derive"]
 
 [workspace.dependencies]
 polars = "0.44.2"
+
+[workspace.package]
+version = "0.1.0"
+edition = "2021"
+license = "MIT"
+repository = "https://github.com/letkemann/polars-row-iter"
