+++ polars-rows-iter-derive/Cargo.toml
@@ -1,7 +1,9 @@
 [package]
 name = "polars-rows-iter-derive"
-version = "0.1.0"
-edition = "2021"
+version.workspace = true
+edition.workspace = true
+license.workspace = true
+repository.workspace = true
 
 [lib]
 proc-macro = true
