+++ polars-rows-iter/src/dataframe_rows_iter_ext.rs
@@ -1,0 +1,18 @@
+use polars::prelude::*;
+
+use crate::FromDataFrameRow;
+
+pub trait DataframeRowsIterExt<'a> {
+    fn rows_iter<T>(&'a self) -> PolarsResult<Box<dyn Iterator<Item = PolarsResult<T>> + 'a>>
+    where
+        T: FromDataFrameRow<'a>;
+}
+
+impl<'a> DataframeRowsIterExt<'a> for DataFrame {
+    fn rows_iter<T>(&'a self) -> PolarsResult<Box<dyn Iterator<Item = PolarsResult<T>> + 'a>>
+    where
+        T: FromDataFrameRow<'a>,
+    {
+        Ok(T::from_dataframe(self)?)
+    }
+}
