+++ polars-rows-iter/src/from_dataframe_row.rs
@@ -1,0 +1,7 @@
+use polars::prelude::*;
+
+pub trait FromDataFrameRow<'a> {
+    fn from_dataframe(dataframe: &'a DataFrame) -> PolarsResult<Box<dyn Iterator<Item = PolarsResult<Self>> + 'a>>
+    where
+        Self: Sized;
+}
